Revert "Merge vdbdd/renderer_v2 into main. (#2098)"
This reverts commit ac5da4815d7a2a84e69ca25a07056cd688f1d604.

+++ crates/lgn-graphics-renderer/src/resources/persistent_descriptor_set_manager.rs
@@ -112,7 +112,7 @@ impl PersistentDescriptorSetManager {
     }
 
     pub fn frame_update(&mut self) {
-        self.render_frame = (self.render_frame + 1) % self.num_render_frames;
+        self.render_frame = (self.render_frame + 1) % self.num_render_frames as u64;
         self.bindless_index_allocator
             .release_indexes(&self.removed_indices[self.render_frame as usize]);
         self.removed_indices[self.render_frame as usize].clear();
